refactor: LottoGame class 분리
